refactor(ai-gateway): split provider definitions
