+++ apps/web/src/lib/ai-gateway/model-api-kinds.ts
@@ -1,5 +1,6 @@
 import { findKiloExclusiveModel } from '@/lib/ai-gateway/models';
-import { OPENROUTER, tryGetProviderById } from '@/lib/ai-gateway/providers/provider-definitions';
+import { OPENROUTER } from '@/lib/ai-gateway/providers/definitions/openrouter';
+import { tryGetProviderById } from '@/lib/ai-gateway/providers/definitions/try-get-provider-by-id';
 import type { GatewayChatApiKind } from '@/lib/ai-gateway/providers/types';
 
 const GATEWAY_CHAT_API_KINDS: readonly GatewayChatApiKind[] = [
