+++ apps/web/src/lib/ai-gateway/providers/provider-definitions.test.ts
@@ -1,10 +1,8 @@
 import { describe, expect, test } from '@jest/globals';
 
-import {
-  LONGCAT,
-  OPENROUTER,
-  tryGetProviderById,
-} from '@/lib/ai-gateway/providers/provider-definitions';
+import { LONGCAT } from '@/lib/ai-gateway/providers/definitions/longcat';
+import { OPENROUTER } from '@/lib/ai-gateway/providers/definitions/openrouter';
+import { tryGetProviderById } from '@/lib/ai-gateway/providers/definitions/try-get-provider-by-id';
 import type { GatewayRequest } from '@/lib/ai-gateway/providers/openrouter/types';
 import {
   ReasoningDetailsTransform,
